Update toggle & delete functions
For the toggle function, using the event (e) prop, the aria-pressed attribute value for the target button is updated to either true or false based on its previous state and only when its pressed.

For the delete function, setFilteredList now filters the filtered list and not the original list like before. This prevents the list from reloading to showing All extensions when deleting extensions when viewing either the Active or Inactive extensions.